requirements.txt
numpy
healpy
scipy
sys
os
ConfigParser
astropy.io
